requirements.txt
# ─────────────────────────────────────────────
# Core API Framework
# ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0

# ─────────────────────────────────────────────
# Database
# ─────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0

# ─────────────────────────────────────────────
# LLM - Groq (current active models April 2026)
# Primary: llama-3.3-70b-versatile
# Fast:    llama-3.1-8b-instant
# ─────────────────────────────────────────────
groq==0.9.0

# ─────────────────────────────────────────────
# Agent Orchestration - LangGraph
# ─────────────────────────────────────────────
langgraph==0.1.5
langchain-core==0.2.3

# ─────────────────────────────────────────────
# Vector Store
# chromadb 0.4.24 = last stable before posthog telemetry bug
# ─────────────────────────────────────────────
chromadb==0.4.24
sentence-transformers==3.0.0

# ─────────────────────────────────────────────
# OCR (local dev)
# ─────────────────────────────────────────────
easyocr==1.7.1
Pillow==10.3.0
pypdf==4.2.0
pdf2image==1.17.0
pytesseract==0.3.10

# ─────────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
tenacity==8.3.0
numpy==1.26.4

# ─────────────────────────────────────────────
# Testing
# ─────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-timeout==2.3.1
pytest-cov==5.0.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
